requirements.txt
frozen-flask~=1.0.2
flask~=3.1.2
flask-bootstrap~=3.3.7.1
python-dotenv~=1.1.1